inventory/requirements.txt
fastapi==0.128.2
uvicorn==0.40.0
redis-om==1.0.6
python-dotenv==1.2.1
